Move decode_csv_field processor to own package
So that it can be included selectively by different Beats.
