+++ libbeat/processors/script/javascript/module/processor/processor.go
@@ -31,6 +31,7 @@ import (
 	"github.com/elastic/beats/libbeat/processors/add_locale"
 	"github.com/elastic/beats/libbeat/processors/add_process_metadata"
 	"github.com/elastic/beats/libbeat/processors/communityid"
+	"github.com/elastic/beats/libbeat/processors/decode_csv_field"
 	"github.com/elastic/beats/libbeat/processors/dissect"
 	"github.com/elastic/beats/libbeat/processors/dns"
 	"github.com/elastic/beats/libbeat/processors/script/javascript"
@@ -48,7 +49,7 @@ var constructors = map[string]processors.Constructor{
 	"AddProcessMetadata":    add_process_metadata.New,
 	"CommunityID":           communityid.New,
 	"CopyFields":            actions.NewCopyFields,
-	"DecodeCSVField":        actions.NewDecodeCSVField,
+	"DecodeCSVField":        decode_csv_field.NewDecodeCSVField,
 	"DecodeJSONFields":      actions.NewDecodeJSONFields,
 	"Dissect":               dissect.NewProcessor,
 	"DNS":                   dns.New,
